requirements.txt
# Core face recognition and computer vision
insightface==0.7.3
onnxruntime==1.22.0
opencv-python==4.8.1.78
numpy==1.24.3

# Image processing
pillow==10.0.1
pillow-avif-plugin==1.4.3

# Web framework
flask==2.3.3
flask-cors==4.0.0

# Additional dependencies for face detection and processing
mediapipe==0.10.8

# Machine learning and deep learning
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0

# Utilities
requests==2.31.0
urllib3==2.0.7
certifi==2023.11.17

# Development and debugging
python-dotenv==1.0.0 